requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-rest-framework==0.1.0
djangorestframework==3.11.2
drf-yasg==1.17.1
h11==0.11.0
httpcore==0.12.0
httpx==0.16.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
rfc3986==1.4.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
